+++ src/x2net/Flows/SingleThreadFlow.cs
@@ -38,7 +38,7 @@ namespace x2net
             this.name = name;
         }
 
-        public override Flow Startup()
+        public override Flow Start()
         {
             lock (syncRoot)
             {
@@ -57,7 +57,7 @@ namespace x2net
             return this;
         }
 
-        public override void Shutdown()
+        public override void Stop()
         {
             lock (syncRoot)
             {