+++ src/x2net/Links/LinkEvents.cs
@@ -498,12 +498,6 @@ namespace x2net
             }
         }
 
-        public override void Deserialize(VerboseDeserializer deserializer)
-        {
-            base.Deserialize(deserializer);
-            deserializer.Read("Data", out data_);
-        }
-
         public override int GetLength(Type targetType, ref bool flag)
         {
             int length = base.GetLength(targetType, ref flag);
@@ -530,18 +524,6 @@ namespace x2net
             {
                 serializer.Write(data_);
             }
-            if (targetType != null && targetType == typeof(HandshakeReq))
-            {
-                flag = false;
-            }
-        }
-
-        public override void Serialize(VerboseSerializer serializer,
-            Type targetType, ref bool flag)
-        {
-            base.Serialize(serializer, targetType, ref flag);
-            if (!flag) { return; }
-            serializer.Write("Data", data_);
             if (targetType != null && targetType == typeof(HandshakeReq))
             {
                 flag = false;
@@ -674,12 +656,6 @@ namespace x2net
             }
         }
 
-        public override void Deserialize(VerboseDeserializer deserializer)
-        {
-            base.Deserialize(deserializer);
-            deserializer.Read("Data", out data_);
-        }
-
         public override int GetLength(Type targetType, ref bool flag)
         {
             int length = base.GetLength(targetType, ref flag);
@@ -706,18 +682,6 @@ namespace x2net
             {
                 serializer.Write(data_);
             }
-            if (targetType != null && targetType == typeof(HandshakeResp))
-            {
-                flag = false;
-            }
-        }
-
-        public override void Serialize(VerboseSerializer serializer,
-            Type targetType, ref bool flag)
-        {
-            base.Serialize(serializer, targetType, ref flag);
-            if (!flag) { return; }
-            serializer.Write("Data", data_);
             if (targetType != null && targetType == typeof(HandshakeResp))
             {
                 flag = false;
@@ -850,12 +814,6 @@ namespace x2net
             }
         }
 
-        public override void Deserialize(VerboseDeserializer deserializer)
-        {
-            base.Deserialize(deserializer);
-            deserializer.Read("Result", out result_);
-        }
-
         public override int GetLength(Type targetType, ref bool flag)
         {
             int length = base.GetLength(targetType, ref flag);
@@ -882,18 +840,6 @@ namespace x2net
             {
                 serializer.Write(result_);
             }
-            if (targetType != null && targetType == typeof(HandshakeAck))
-            {
-                flag = false;
-            }
-        }
-
-        public override void Serialize(VerboseSerializer serializer,
-            Type targetType, ref bool flag)
-        {
-            base.Serialize(serializer, targetType, ref flag);
-            if (!flag) { return; }
-            serializer.Write("Result", result_);
             if (targetType != null && targetType == typeof(HandshakeAck))
             {
                 flag = false;